feat(parse5): change parser/tokenizer methods to be protected
Some projects in the wild depend on extending `Parser` and `Tokenizer`
by overriding certain methods and using internals we don't publish on
our public API.

This just moves all the private members to be protected so those
projects can upgrade to 7.x.

Not everything should be protected but it seems our current class
members all fit the criteria.